Skip serialisation of default value offsets
Previously, offsets with default values would be serialised when
converting a manifest to JSON. Redpanda's behaviour is different: if an
offset has a default value (i64::MIN) it does not get serialised.

This difference is problematic because Redpanda deserialises these
offsets as unsigned integers, which fails if provided with a negative
value.

This commit fixes the issue by mirroring Redpanda's behaviour and
ignoring offset fields with default values.